removed bankAccountSuccess page
