+++ client/src/pages/Landing.js
@@ -40,6 +40,8 @@ function LandingPage(props) {
         return <PaymentPage />;
       case "success":
         return <SuccessPage />;
+      case "bank-account/success":
+        return <SuccessPage />;
       case "sitter-profile":
         return <SitterProfile profileId={"5e86446fb63eabc010ee2828"} />;
       case "/":
